refactor: `ServiceDetail`에서 `Carousel` 분리
